Bump numpy to 1.23.4
Previously, NumPy was pinned to an old version on Windows to avoid a bug
in a Windows update.  However, that bug has now been fixed and that old
NumPy version is now incompatible with the version of Python now in the
Windows store (Python 3.10).  The old version also has known security
vulnerabilities, giving one final reason to update.